pyproject.toml
# ============================================================
# DMD+SMV Tactile Internet - Project Configuration
# Author: Ali Vahedi (Mohammad Ali Vahedifar)
# IEEE INFOCOM 2025
# ============================================================

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dmd-smv-tactile-internet"
version = "1.0.0"
description = "Discrete Mode Decomposition meets Shapley Value for Robust Signal Prediction in Tactile Internet"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Ali Vahedi (Mohammad Ali Vahedifar)", email = "av@ece.au.dk"},
    {name = "Qi Zhang", email = "qz@ece.au.dk"},
]
maintainers = [
    {name = "Ali Vahedi", email = "av@ece.au.dk"},
]
keywords = [
    "tactile internet",
    "haptic signals",
    "discrete mode decomposition",
    "shapley value",
    "signal prediction",
    "deep learning",
    "transformer",
    "teleoperation",
    "6G",
    "edge intelligence",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Signal Processing",
]
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "torch>=2.0.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
    "loguru>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
]
viz = [
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
]
all = [
    "dmd-smv-tactile-internet[dev,docs,viz]",
]

[project.urls]
Homepage = "https://github.com/alivahedi/dmd-smv-tactile-internet"
Documentation = "https://github.com/alivahedi/dmd-smv-tactile-internet/docs"
Repository = "https://github.com/alivahedi/dmd-smv-tactile-internet"
Issues = "https://github.com/alivahedi/dmd-smv-tactile-internet/issues"

[project.scripts]
dmd-train = "experiments.train_all_models:main"
dmd-eval = "experiments.evaluate_accuracy:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/setup.py"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = [
    "tests",
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
